Move the single function of services.go to driver.go
