--- pkg/controller/kibana/services.go
@@ -1,33 +1,0 @@
-// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
-// or more contributor license agreements. Licensed under the Elastic License;
-// you may not use this file except in compliance with the Elastic License.
-
-package kibana
-
-import (
-	corev1 "k8s.io/api/core/v1"
-
-	kbv1 "github.com/elastic/cloud-on-k8s/pkg/apis/kibana/v1"
-	"github.com/elastic/cloud-on-k8s/pkg/controller/common/defaults"
-)
-
-func NewService(kb kbv1.Kibana) *corev1.Service {
-	svc := corev1.Service{
-		ObjectMeta: kb.Spec.HTTP.Service.ObjectMeta,
-		Spec:       kb.Spec.HTTP.Service.Spec,
-	}
-
-	svc.ObjectMeta.Namespace = kb.Namespace
-	svc.ObjectMeta.Name = HTTPService(kb.Name)
-
-	labels := NewLabels(kb.Name)
-	ports := []corev1.ServicePort{
-		{
-			Name:     kb.Spec.HTTP.Protocol(),
-			Protocol: corev1.ProtocolTCP,
-			Port:     HTTPPort,
-		},
-	}
-
-	return defaults.SetServiceDefaults(&svc, labels, labels, ports)
-}
